fattureGET

+++ gestionale_anziendale_energia/src/main/java/com/buildweek/gestionale_anziendale_energia/models/Fattura.java
@@ -37,7 +37,7 @@ public class Fattura {
 	@Column(nullable = false)
 	private LocalDate dataFattura;
 	@Column(nullable = false)
-	private double importo;
+	private Long importo;
 	@Enumerated(EnumType.STRING)
 	private StatoFattura statoFattura;
